impl custom flow 2.0 on multiple pages in a single playwright session
